Fix done button status bar bug
When the status bar is hidden, its height property is set
to zero, this makes it difficult to position elements relative
to that without them moving when its hidden. To fix this I've
hardcoded the value to twenty. There wasn't an easy way of getting
this number without displaying the status bar